Updated styling and better structure
